chore: Remove `rollup-plugin-visualizer` dependency and configuration, and add `React` import with `prop-types` linting disabled.
