+++ package.json
@@ -48,7 +48,6 @@
     "knip": "^5.66.2",
     "postcss": "^8.5.3",
     "prettier": "^3.6.2",
-    "rollup-plugin-visualizer": "^5.14.0",
     "tailwindcss": "^4.1.3",
     "vite": "^6.4.1",
     "wrangler": "^4.8.0"
