+++ vite.config.js
@@ -2,19 +2,9 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import path from "path";
 import tailwindcss from "@tailwindcss/vite";
-import { visualizer } from "rollup-plugin-visualizer";
 
 export default defineConfig({
-  plugins: [
-    react(),
-    tailwindcss(),
-    visualizer({
-      open: true, // opens report in browser after build
-      gzipSize: true,
-      brotliSize: true,
-      filename: "bundle-report.html",
-    }),
-  ],
+  plugins: [react(), tailwindcss()],
   resolve: {
     alias: {
       // eslint-disable-next-line no-undef
